src/requirements.txt
cffi==1.9.1
cloudinary==1.5.0
cryptography==1.6
dj-database-url==0.4.1
Django==1.10.4
django-cors-headers==2.0.0
django-rest-knox==2.2.0
django-webpack-loader==0.4.1
djangorestframework==3.5.3
djangorestframework-timed-auth-token==1.2.0
factory-boy==2.8.1
Faker==0.7.10
gunicorn==19.6.0
httpie==0.9.6
idna==2.1
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.17
Pygments==2.1.3
pyOpenSSL==16.2.0
python-dateutil==2.6.0
requests==2.12.3
six==1.10.0
urllib3==1.19.1
whitenoise==3.2.2
